Moves handler related classes to handler package
